Split logic up into different files.
Instead of having Basket & Products in
the same file, we now have different files for it.
